requirements.txt
requests
bs4
tweepy
textblob
matplotlib.pyplot
pandas
sklearn
numpy